Organize the code in client/cas.go
[`go/pkg/client/cas.go`](https://github.com/bazelbuild/remote-apis-sdks/blob/e62cf9b8696a2316c50ffe69bafe0a7972bcddff/go/pkg/client/cas.go)
was approaching the 2k lines mark with 95 symbols. Split the file into two, one
for upload behaviour and one for download behaviour.

This PR has no functional changes.

Tested by successfully running `bazel build //...`.

This is a pre-work for merging in
[`go/pkg/cas/`](https://github.com/bazelbuild/remote-apis-sdks/tree/e62cf9b8696a2316c50ffe69bafe0a7972bcddff/go/pkg/cas).
